sorted the league xp board by lifetime xp with new firebase variable

+++ PerfectPronunciation/Model/League.swift
@@ -10,7 +10,7 @@ struct League: Identifiable {
     var id : String
     var userName : String
     var country : String
-    var experience : Double
+    var experience : Int
     var league : String
 }
